pyproject.toml
[tool.poetry]
name = "sublime_music"
version = "0.11.9"
description = "A native GTK *sonic client."
license = "GPL-3.0-or-later"
authors = ["Sumner Evans <inquiries@sumnerevans.com>"]
readme = "README.rst"
homepage = "https://sublimemusic.app"
repository = "https://sr.ht/~sumner/sublime-music"
documentation = "https://sublime-music.gitlab.io/sublime-music/"
keywords = ["airsonic", "music", "chromecast", "subsonic"]
classifiers = [
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX",
]

# TODO
exclude = ["tests"]

[tool.poetry.urls]
"Bug Tracker" = "https://todo.sr.ht/~sumner/sublime-music"

[tool.poetry.scripts]
sublime-music = 'sublime_music.__main__:main'

[tool.poetry.dependencies]
python = "^3.8"
bleach = "^3.2.1"
dataclasses-json = "^0.5.2"
deepdiff = "^5.0.2"
fuzzywuzzy = "^0.18.0"
peewee = "^3.13.3"
PyGObject = "^3.38.0"
python-dateutil = "^2.8.1"
python-Levenshtein = "^0.12.0"
python-mpv = "^0.5.2"
requests = "^2.24.0"
semver = "^2.10.2"
bottle = {version = "^0.12.18", optional = true}
keyring = {version = "^21.4.0", optional = true}
pychromecast = {version = "^7.3.0", optional = true}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
docutils = "^0.16"
flake8 = "^3.8.3"
flake8-annotations = "^2.4.0"
flake8-bugbear = "^20.1.4"
flake8-comprehensions = "^3.2.3"
flake8-import-order = "^0.18.1"
flake8-pep3101 = "^1.3.0"
flake8-print = "^3.1.4"
mypy = "^0.782"
pytest-cov = "^2.10.1"
termcolor = "^1.1.0"
requirements-parser = "^0.2.0"

[tool.poetry.extras]
chromecast = ["pychromecast"]
keyring = ["keyring"]
server = ["bottle"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
exclude = '''
(
  /flatpak/
)
'''
